+++ src/Foldable.ts
@@ -81,131 +81,32 @@ export interface Foldable4<F extends URIS4> {
 /**
  * @since 3.0.0
  */
-export interface FoldableComposition<F, G> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => (fga: HKT<F, HKT<G, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => (fga: HKT<F, HKT<G, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => (fga: HKT<F, HKT<G, A>>) => B
+export const reduceComposition: <F, G>(
+  F: Foldable<F>,
+  G: Foldable<G>
+) => <A, B>(b: B, f: (b: B, a: A) => B) => (fga: HKT<F, HKT<G, A>>) => B = (F, G) => (b, f) =>
+  F.reduce(b, (b, ga) => pipe(ga, G.reduce(b, f)))
+
+/**
+ * @since 3.0.0
+ */
+export const foldMapComposition: <F, G>(
+  F: Foldable<F>,
+  G: Foldable<G>
+) => <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => (fga: HKT<F, HKT<G, A>>) => M = (F, G) => (M) => {
+  const foldMapF = F.foldMap(M)
+  const foldMapG = G.foldMap(M)
+  return (f) => foldMapF(foldMapG(f))
 }
 
 /**
  * @since 3.0.0
  */
-export interface FoldableComposition11<F extends URIS, G extends URIS> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => (fga: Kind<F, Kind<G, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => (fga: Kind<F, Kind<G, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => (fga: Kind<F, Kind<G, A>>) => B
-}
-
-/**
- * @since 3.0.0
- */
-export interface FoldableComposition12<F extends URIS, G extends URIS2> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => <E>(fga: Kind<F, Kind2<G, E, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => <E>(fga: Kind<F, Kind2<G, E, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => <E>(fga: Kind<F, Kind2<G, E, A>>) => B
-}
-
-/**
- * @since 3.0.0
- */
-export interface FoldableComposition12C<F extends URIS, G extends URIS2, E> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => (fga: Kind<F, Kind2<G, E, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => (fga: Kind<F, Kind2<G, E, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => (fga: Kind<F, Kind2<G, E, A>>) => B
-}
-
-/**
- * @since 3.0.0
- */
-export interface FoldableComposition21<F extends URIS2, G extends URIS> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => <E>(fga: Kind2<F, E, Kind<G, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => <E>(fga: Kind2<F, E, Kind<G, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => <E>(fga: Kind2<F, E, Kind<G, A>>) => B
-}
-
-/**
- * @since 3.0.0
- */
-export interface FoldableComposition2C1<F extends URIS2, G extends URIS, E> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => (fga: Kind2<F, E, Kind<G, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => (fga: Kind2<F, E, Kind<G, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => (fga: Kind2<F, E, Kind<G, A>>) => B
-}
-
-/**
- * @since 3.0.0
- */
-export interface FoldableComposition22<F extends URIS2, G extends URIS2> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => <D, E>(fga: Kind2<F, D, Kind2<G, E, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => <D, E>(fga: Kind2<F, D, Kind2<G, E, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => <D, E>(fga: Kind2<F, D, Kind2<G, E, A>>) => B
-}
-
-/**
- * @since 3.0.0
- */
-export interface FoldableComposition22C<F extends URIS2, G extends URIS2, E> {
-  readonly reduce: <A, B>(b: B, f: (b: B, a: A) => B) => <D>(fga: Kind2<F, D, Kind2<G, E, A>>) => B
-  readonly foldMap: <M>(M: Monoid<M>) => <A>(f: (a: A) => M) => <D>(fga: Kind2<F, D, Kind2<G, E, A>>) => M
-  readonly reduceRight: <A, B>(b: B, f: (a: A, b: B) => B) => <D>(fga: Kind2<F, D, Kind2<G, E, A>>) => B
-}
-
-/**
- * Returns the composition of two foldables
- *
- * @example
- * import { getFoldableComposition } from 'fp-ts/lib/Foldable'
- * import { foldableReadonlyArray } from 'fp-ts/lib/ReadonlyArray'
- * import { foldableOption, some, none } from 'fp-ts/lib/Option'
- * import { monoidString } from 'fp-ts/lib/Monoid'
- * import { pipe } from 'fp-ts/lib/function'
- *
- * const F = getFoldableComposition(foldableReadonlyArray, foldableOption)
- * assert.strictEqual(pipe([some('a'), some('b'), some('c')], F.reduce('', monoidString.concat)), 'abc')
- * assert.strictEqual(pipe([some('a'), none, some('c')], F.reduce('', monoidString.concat)), 'ac')
- *
- * @since 3.0.0
- */
-export function getFoldableComposition<F extends URIS2, G extends URIS2, E>(
-  F: Foldable2<F>,
-  G: Foldable2C<G, E>
-): FoldableComposition22C<F, G, E>
-export function getFoldableComposition<F extends URIS2, G extends URIS2>(
-  F: Foldable2<F>,
-  G: Foldable2<G>
-): FoldableComposition22<F, G>
-export function getFoldableComposition<F extends URIS2, G extends URIS, E>(
-  F: Foldable2C<F, E>,
-  G: Foldable1<G>
-): FoldableComposition2C1<F, G, E>
-export function getFoldableComposition<F extends URIS2, G extends URIS>(
-  F: Foldable2<F>,
-  G: Foldable1<G>
-): FoldableComposition21<F, G>
-export function getFoldableComposition<F extends URIS, G extends URIS2, E>(
-  F: Foldable1<F>,
-  G: Foldable2C<G, E>
-): FoldableComposition12C<F, G, E>
-export function getFoldableComposition<F extends URIS, G extends URIS2>(
-  F: Foldable1<F>,
-  G: Foldable2<G>
-): FoldableComposition12<F, G>
-export function getFoldableComposition<F extends URIS, G extends URIS>(
-  F: Foldable1<F>,
-  G: Foldable1<G>
-): FoldableComposition11<F, G>
-export function getFoldableComposition<F, G>(F: Foldable<F>, G: Foldable<G>): FoldableComposition<F, G>
-export function getFoldableComposition<F, G>(F: Foldable<F>, G: Foldable<G>): FoldableComposition<F, G> {
-  return {
-    reduce: (b, f) => F.reduce(b, (b, ga) => pipe(ga, G.reduce(b, f))),
-    foldMap: (M) => {
-      const foldMapF = F.foldMap(M)
-      const foldMapG = G.foldMap(M)
-      return (f) => foldMapF(foldMapG(f))
-    },
-    reduceRight: (b, f) => F.reduceRight(b, (ga, b) => pipe(ga, G.reduceRight(b, f)))
-  }
-}
+export const reduceRightComposition: <F, G>(
+  F: Foldable<F>,
+  G: Foldable<G>
+) => <A, B>(b: B, f: (a: A, b: B) => B) => (fga: HKT<F, HKT<G, A>>) => B = (F, G) => (b, f) =>
+  F.reduceRight(b, (ga, b) => pipe(ga, G.reduceRight(b, f)))
 
 /**
  * Similar to 'reduce', but the result is encapsulated in a monad.